requirements.txt
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
lxml==5.4.0
pydub==0.25.1
PyMuPDF==1.26.0
python-docx==1.1.2
requests==2.32.3
soupsieve==2.7
typing_extensions==4.13.2
urllib3==2.4.0
wget==3.2
